Begun the long process of upgrading the game to use HD graphics and 16:9 aspect ratio. Many things will be out of place, but functional nonetheless. Exposed missing Size metrics for many graphic elements (still a work in progress) Upgraded and reenabled the Resolution option in the Options Screen. Upgraded Menu objects to handle sound events themselves. No more menu sounds in inappropriate places! Added MENU_INVALID_DECIDE sound event (handled by Menus)

+++ trunk/WindowsGame1/WindowsGame1/Drawing/BpmMeter.cs
@@ -192,7 +192,7 @@ namespace WGiBeat.Drawing
             var meterBPM = Math.Max(BpmLevels[BpmLevels.Count() - 1], DisplayedSong.StartBPM * (1 - beatFraction));
 
           
-            int height = (this.Height - 2) / _meterSprite.Rows;
+            float height = (this.Height - 2) / _meterSprite.Rows;
             for (int x = 0; x < BpmLevels.Count(); x++)
             {
                 if ((meterBPM >= BpmLevels[x]))
